Generate chart of a twitter users sentiments
Complete the implementation of search in application.py in such a way that the function:
- queries Twitter’s API for a user’s most recent 100 tweets,
- classifies each tweet as positive, negative, or neutral,
- generates a chart that accurately depicts those sentiments as percentages.

Steps required to run the twitter app
https://apps.twitter.com/app/14376222/keys

Click Keys and Access Tokens again.

Highlight and copy the value to the right of Consumer Key (API Key).

In a terminal window, execute

export API_KEY=value

where value is that (pasted) value, without any space immediately before or after the =.

Highlight and copy the value to the right of Consumer Secret (API Secret).

In a terminal window, execute

export API_SECRET=value

where value is that (pasted) value, without any space immediately before or after the =.