Allow a user to register for an account.
Require that a user input a username. Render an apology if the user’s input is blank or the username already exists.

Require that a user input a password and then that same password again. Render an apology if either input is blank or the passwords do not match.

INSERT the new user into users, storing a hash of the user’s password, not the password itself. Use https://passlib.readthedocs.io/en/stable/lib/passlib.context.html#passlib.context.CryptContext.hash to hash passwords

Create a new template (register.html)

Users should be able to register for an account and log in