Buy stock for user
Require that a user input a stock’s symbol. Render an apology if the input is blank or the symbol does not exist (as per the return value of lookup).

Require that a user input a number of shares. Render an apology if the input is not a positive integer.

Add portfolio table to finance.db via which to keep track of the purchase. Store enough information so that you know who bought what at what price.
- Define UNIQUE index on the “portfolio.id” field
- Define (non-UNIQUE) indexes on any fields via which to search (as via SELECT with WHERE).