Fix bug (homepage): Display correct number of rows
The previous code used a query which returned the number of stock purchases
made for the first (alphabetic) stock in a users portfolio.

The code was originally believed to return the number of distinct symbols
in a users portfolio. Based on this mis-understanding, the code
used this number to determine the number of rows to display in the
table on the homepage.

The number has been updated to accurately represent the number of unique
stocks owned by the user.