Convert div to img element.
Originally this was a div with the class .product-image.

However, to set the image src attribute, I needed to select the child node.

Rather than arbitrarily select the first child node of the div, which could have lead to bugs down the road, I applied the .product-image class directly to the img element.