Call functions at top of script for readability.
The functions execute in an improper order. The process to receive the edit form, then submit edits needs to be done in series.

To achieve this I will use a promise constructor then chain together the various promises.

The "TODO" note is included in this commit.

The function calls have also been moved to the $( document ).ready() function at the top of the script file. This is to enhance readbility and to make it clear what task happen in which order.