Parse incoming POST request data into a dictionary
The Fetch API is used to send a POST request from the client.

Our server gets the incoming JSON using the werkzeug API method, .get_json().

The json is put into a variable of type 'list'.

The list needs to be converted to a dictionary with corresponding key value pairs.
This affords us the ability to access the relevant data for our database query.

The code here pulls product data from our database with the corresponding id.
However, it does not update the database with the users requests.
(This functionality is in the commented out code. It will come in another commit.)