Send POST request with product edits
Construct promises for each function to give more control over when functions execute.

Send a POST request to the server.
Resolve if a json is returned.

Add TODO items.

Listen for multiple clicks on the save button, rather than a single click.
This is primarily for testing the POST request.

TODO: Still need to listen for the correct click events at the correct time.