Send reference ids with product information.
This will be needed when we want to determine which reference to edit later on.

This change is part of the following bug fix.

Bug: References not correctly organized for POST request
Right now I only send the Reference link change to the survey and not in a logical format.
Fix:
Label each reference list item with the reference id
Send the id from the reference table to the document.
Set the id attribute value for the list item to the id for that reference.
Name the form, or set the form id to be reference-form-[id] so we can find it later.
Create an array specifically for references.
For each reference form
Create an array
Store the title and link.
Append this title and link to the references array
Append the references array to the requested edits Map
Turn the map into an array
Make sure this array has access to the array of references
Convert the array to a string
Make sure no reference data was lost
Send the string with a JSON header
Confirm the server received all the reference information.
Is this reference information tiered or flat with the other data?
Store the reference data in an easy-to-access way in our dictionary.
Make updates to the product and reference tables.
Return the new product data to the client.