Toggle between edit and view modes
Bug:
1. Click edit
2. Save edits
Expected: An edit button to appear at the bottom of the page.
Actual: A save button still appeared at the bottom of the page.

Each time the product enters view-mode, then the save button is removed and the edit button is rendered.

Each time the product enters edit-mode, then the edit button is removed and the save button is rendered.